refactor(game): extract turnFlow.ts (Phase 5 of #132)
Move the turn-progression logic into src/Game/turnFlow.ts as free
functions: roll, switchDice, move, moveAndFinalize, toMoved,
executeAndRecalculate, checkAndCompleteTurn, confirmTurn, startMove,
and getPlayersForColor (used by roll). The public Game.* statics now
delegate via 'public static X = X' to the imported free functions, so
external callers (ai/api/client) and their type signatures are
unchanged. Internal cross-calls go module-local (Rule A):
roll->getPlayersForColor, moveAndFinalize->move/checkAndCompleteTurn,
executeAndRecalculate->move/checkAndCompleteTurn/confirmTurn,
checkAndCompleteTurn->toMoved.

The player accessors activePlayer/inactivePlayer and the robot helpers
(handleRobotMovedState, executeRobotTurn) stay on the class;
handleRobotMovedState still reaches confirmTurn through the facade.

Byte-exact relocation (mechanical extraction script) preserves behavior;
the test suite is identical before and after. Removed now-dead value
imports (Dice, Play, MoveExecutionOptions) from index.ts. index.ts drops
~1500 lines (~2050 -> 586).

Note: switchDice coverage (weak, ~#134) is left as a follow-up coverage
pass; this phase is a behavior-preserving move.

No behavior change. Core: 475 passed / 12 skipped, tsc -b clean.

Refs #132
